+++ parray.h
@@ -1,1 +1,1 @@
-unsigned int parray ( float ** x, unsigned int n, unsigned int m, float z, unsigned int * P );
+unsigned int parray ( double ** x, unsigned int n, unsigned int m, double z, unsigned int * P );
